backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.3.0
numpy==1.26.4
scipy==1.13.0
requests==2.32.3
python-jose==3.3.0
aiofiles==23.2.1
slowapi==0.1.9
loguru==0.7.2
exifread==3.0.0
